changed to Float32

+++ src/SheetModel.jl
@@ -7,9 +7,9 @@ using ParallelStencil #, CUDA
 #CUDA.allowscalar(false) # don't allow scalar indexing which is not GPU compatible
 #using ParallelStencil.FiniteDifferences2D
 @static if USE_GPU
-    @init_parallel_stencil(CUDA, Float64, 2)
+    @init_parallel_stencil(CUDA, Float32, 2)
 else
-    @init_parallel_stencil(Threads, Float64, 2)
+    @init_parallel_stencil(Threads, Float32, 2)
 end
